Extract knowledge indexing identity into indexing_config.py
Move IndexingSettings, the storage-key helpers (now public
storage_key_for_base / indexing_settings_key), and the pure
chroma_collection_exists probe out of manager.py into a focused
indexing_config module. registry.py now imports these from their
real homes (plus ChromaDb, Knowledge, and create_configured_embedder
directly) instead of reaching into manager privates, so registry no
longer imports manager at all. Storage keys, collection names, and
persisted metadata stay byte-identical.
